move pingEndpoint to utils
